Add files via upload
Computer players now have a full strategy for picking a card to play. Obviously it could be greatly improved. Currently it only bases decisions on the cards in the current trick and in it's hand, not on what it 'remembers' about previous tricks, nor on the game score.
It still just passes the 3 right-most cards in it's hand; to do next.

The print streams are still not done being organized. Also need to figure out if there should be a single game print stream, or if the game should be printing to all player print streams. And just making the output nicer in general.

Finally, still need to have the main method ask human for (a) name(s) and then setup the game.