Add files via upload
Moved Round as an inner class of Game.
Gave Game a PrintStream.

Still need to make the printing/displaying better, mostly with respect to newlines. The ideal situation is that the user has two windows, one for their own personal print stream and one for the game print stream (which everyone sees). But they may also have these in the same window. Want output to reasonable in both scenarios.

From before:
Still need to have the main method ask human for (a) name(s) and then setup the game.